Move utils.py to project folder.
Since `passes_pydantic_multiple_validator` is used in the package internals we either have to include tests when building the package or move the utils.py. The latter seems like the better option.
