+++ tests/constraints/test_float_constraints.py
@@ -8,7 +8,7 @@ from pydantic import ConstrainedFloat
 from pydantic_factories.constraints.constrained_float_handler import (
     handle_constrained_float,
 )
-from tests.utils import passes_pydantic_multiple_validator
+from pydantic_factories.utils import passes_pydantic_multiple_validator
 
 
 def create_constrained_field(
